refactor: prio tests changed to be generic traffic tests
Two types of peers can be defined now: normal peers which generate and receive some traffic that is measured and
high traffic peers which try to flood the node with some other traffic.

Some of the conecpts need to be extended - factory should contain also first message generation (as well as custom
message with some custom values) and each payload type should contain matcher associated with it.